pow is total overkill for x*x, introduce pow2 inline function to fix it. fix a warning.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2323 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/PluginManager.cpp
@@ -257,7 +257,7 @@ CPluginInfo::CPluginInfo(const char *_rFilename)
    If the user has done that he will instead get the "Can't open %s, it's missing" message. */
 void CPluginManager::GetPluginInfo(CPluginInfo *&info, std::string Filename)
 {
-	for (int i = 0; i < m_PluginInfos.size(); i++)
+	for (int i = 0; i < (int)m_PluginInfos.size(); i++)
 	{
 		if (m_PluginInfos.at(i).GetFilename() == Filename)
 		{
